+++ src/e2e.test.ts
@@ -3390,6 +3390,9 @@ async function serve(
     stdout(s) {
       output += s
     },
+    stderr(s) {
+      output += s
+    },
     exit(code) {
       exitCode = code
     },