+++ src/plugins/ConnectRpc.test.ts
@@ -14,6 +14,9 @@ async function serve(
     stdout(s) {
       output += s
     },
+    stderr(s) {
+      output += s
+    },
     exit(code) {
       exitCode = code
     },